minor: move components into separate folder
