+++ client/src/lib/components/job-post-creation/FormSidebar.svelte
