+++ client/src/lib/components/job-post-creation/PreviewDrawer.svelte
