+++ client/src/routes/company/edit/[id]/+page.svelte
@@ -1,12 +1,12 @@
 <script>
   import { goto } from '$app/navigation';
   import { ChevronLeft } from 'lucide-svelte';
-  import BasicInfoForm from '$lib/components/forms/job-post-creation/BasicInfoForm.svelte';
-  import DescriptionForm from '$lib/components/forms/job-post-creation/DescriptionForm.svelte';
-  import RequirementForm from '$lib/components/forms/job-post-creation/RequirementForm.svelte';
-  import PostSettingForm from '$lib/components/forms/job-post-creation/PostSettingForm.svelte';
-  import PreviewDrawer from '$lib/components/PreviewDrawer.svelte';
-  import FormSidebar from '$lib/components/FormSidebar.svelte';
+  import BasicInfoForm from '$lib/components/job-post-creation/forms/BasicInfoForm.svelte';
+  import DescriptionForm from '$lib/components/job-post-creation/forms/DescriptionForm.svelte';
+  import RequirementForm from '$lib/components/job-post-creation/forms/RequirementForm.svelte';
+  import PostSettingForm from '$lib/components/job-post-creation/forms/PostSettingForm.svelte';
+  import PreviewDrawer from '$lib/components/job-post-creation/PreviewDrawer.svelte';
+  import FormSidebar from '$lib/components/job-post-creation/FormSidebar.svelte';
   
   let isPreviewOpen = $state(false);
   let activeSection = $state('basic-info');
